+++ src-tauri/src/manager/downloader.rs
@@ -350,7 +350,7 @@ impl<'a> Installer<'a> {
         self.check_cancelled()?;
         self.update(InstallTask::Extracting);
 
-        zip_extract::extract(Cursor::new(data), &path, false)
+        util::zip::extract(Cursor::new(data), &path)
             .fs_context("extracting mod", &path)?;
 
         normalize_mod_structure(&mut path)?;
@@ -440,7 +440,7 @@ impl<'a> Installer<'a> {
 pub fn normalize_mod_structure(path: &mut PathBuf) -> Result<()> {
     for dir in ["BepInExPack", "BepInEx", "plugins"].iter() {
         path.push(dir);
-        util::io::flatten_if_exists(&*path)?;
+        util::fs::flatten_if_exists(&*path)?;
         path.pop();
     }
 
@@ -533,7 +533,7 @@ fn install_from_disk_default(src: &Path, dest: &Path, name: &str) -> Result<()> 
             if entry_name == "config" {
                 let target_path = target_path.join("config");
                 fs::create_dir_all(&target_path)?;
-                util::io::copy_contents(&entry_path, &target_path, false)
+                util::fs::copy_contents(&entry_path, &target_path, false)
                     .fs_context("copying config", &entry_path)?;
             } else {
                 let target_path = match entry_name.to_string_lossy().as_ref() {
@@ -543,7 +543,7 @@ fn install_from_disk_default(src: &Path, dest: &Path, name: &str) -> Result<()> 
                 };
 
                 fs::create_dir_all(target_path.parent().unwrap())?;
-                util::io::copy_dir(&entry_path, &target_path, false)
+                util::fs::copy_dir(&entry_path, &target_path, false)
                     .fs_context("copying directory", &entry_path)?;
             }
         } else {
@@ -561,13 +561,13 @@ fn install_from_disk_bepinex(src: &Path, dest: &Path) -> Result<()> {
     // Some BepInEx packs come with a subfolder where the actual BepInEx files are
     for entry in fs::read_dir(src)? {
         let entry_path = entry?.path();
-        let entry_name = util::io::file_name(&entry_path);
+        let entry_name = util::fs::file_name(&entry_path);
 
         if entry_path.is_dir() && entry_name.contains("BepInEx") {
             // ... and some have even more subfolders ...
             // do this first, since otherwise entry_path will be removed already
-            util::io::flatten_if_exists(&entry_path.join("BepInEx"))?;
-            util::io::flatten_if_exists(&entry_path)?;
+            util::fs::flatten_if_exists(&entry_path.join("BepInEx"))?;
+            util::fs::flatten_if_exists(&entry_path)?;
         }
     }
 
@@ -579,7 +579,7 @@ fn install_from_disk_bepinex(src: &Path, dest: &Path) -> Result<()> {
             let target_path = target_path.join(entry_name);
             fs::create_dir_all(&target_path)?;
 
-            util::io::copy_contents(&entry_path, &target_path, false)
+            util::fs::copy_contents(&entry_path, &target_path, false)
                 .fs_context("copying directory", &entry_path)?;
         } else if ["winhttp.dll", ".doorstop_version"]
             .into_iter()