+++ src-tauri/src/util/mod.rs
@@ -1,4 +1,5 @@
 pub mod cmd;
 pub mod error;
-pub mod io;
+pub mod fs;
 pub mod window;
+pub mod zip;